11.5 新增上传图片到七牛云服务器,新增更新修改查重 11.5 新增发送阿里云短信接口,签名和模板尚未配置,还不可用 11.5 新增redis缓存接口 11.12修复电话号码bug

+++ src/main/java/com/theoldsweb/myweb/common/dto/userDto.java
@@ -2,6 +2,7 @@ package com.theoldsweb.myweb.common.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Data
@@ -18,7 +19,7 @@ public class userDto  implements Serializable {
     private String userPassword;
 
 
-    private Integer userPhoneNumber;
+    private BigInteger userPhoneNumber;
 
     private Date createTime;
